Decode into a IWICBitmap instance
Use the technique of the AIT codec sample and decode into an IWICBitmap instance. This makes it possible to just forward the calls to the IWICBitmap instance.
Also move the timing when to decode to when GetFrame is called. This makes it possible to cache the created jpegls_bitmap_frame_decode and provide support for many CopyPixel calls
To display the image, it needs to be converted to GUID_WICPixelFormat32bppPBGRA. The WIC converter code will call CopyPixel line for line.