Introduce SPDX Unique License Identifiers (MIT)
Apply the modern way of documenting the used license, by referencing a SPDX Unique License Identifier in the source code files.

Note: From the legal point of view, this commit is only a change to the textual representation of the license information, but in no way any change to the actual license terms. With this commit applied, all files will still be licensed under the same terms they were before.